perf: split the SDK-linked repositories out of setupRepositories (WALLET-1421)
The previous commit took casper-js-sdk off the account-hash and network
constant paths, but a wallet client still linked it the moment it built
its repositories: setupRepositories() constructs all ten in one call, and
TxSignatureRequestRepository, EIP712Repository and ValidatorsRepository
each reach the SDK. No amount of tree shaking on the client side can
separate constructions that happen in the same function.

Split the factory:
- setupData.ts — setupDataRepositories(), the eight repositories a home
  screen renders from. It also returns the httpDataProvider and the
  resolved logger, so the signing half can share them.
- setupSigning.ts — setupSigningRepositories(), txSignatureRequest and
  eip712, taking those shared pieces as parameters.
- setup.ts — setupRepositories() now composes both. Its return shape and
  parameters are unchanged, so no consumer breaks; it links the SDK, as
  it always did.

Validators:
- data/dto/validators.ts imported DEFAULT_MINIMUM_DELEGATION_AMOUNT and
  DEFAULT_MAXIMUM_DELEGATION_AMOUNT from the SDK — ~900 KB for two
  numbers. Declared locally; validators.test.ts pins them against the
  SDK's own values, so an upstream change fails the build here.

Barrels:
- utils/casperSdk/index.ts and utils/eip712/index.ts no longer re-export
  cep-nft-transfer and eip712/sign. Both barrels are reached from the
  utils barrel that most of src/data imports, so those two re-exports
  made every DTO a transitive SDK importer whenever the host build does
  not tree-shake. They are re-exported from the package root instead —
  the public API is byte-for-byte what it was.
- The SDK-free repositories and DTOs now import specific modules rather
  than the dto/utils/repositories barrels, and the API response types
  they pull from ../repositories are import type.

Regression guard:
- sdk-free-modules.test.ts covers src/setupData.ts, so a value import
  that reaches the SDK from any of the eight repositories fails the test
  run rather than showing up as bundle size in another repo.
