+++ index.ts
@@ -2,3 +2,10 @@ export * from './src/utils';
 export * from './src/setup';
 export * from './src/domain';
 export * from './src/typings';
+// The two SDK-backed util modules. They are re-exported here rather than from the `utils` /
+// `casperSdk` / `eip712` barrels, because those are reached from most of `src/data` — a
+// re-export there made every DTO a transitive `casper-js-sdk` importer. The package root already
+// links the SDK through `./src/setup`, so nothing is lost by exporting them here, and the public
+// API is exactly what it was (WALLET-1421).
+export * from './src/utils/casperSdk/cep-nft-transfer';
+export * from './src/utils/eip712/sign';
